Removed sampling tests within the promise chain, added basic tests for the evaluator, reformatted some old tests, changed NORMAL to LOGNORMAL to reflect the new default.
