+++ src/lib/guesstimator/formatter/formatters/tests/DistributionTextUpToAlternate-test.js
@@ -17,7 +17,8 @@ describe('DistributionTextUpTo', () => {
 
   describe('#format', () => {
     const examples = [
-      [{text: '[8, 10]'}, {guesstimateType: 'NORMAL', low: 8, high: 10}],
+      [{text: '[8, 10]'}, {guesstimateType: 'LOGNORMAL', low: 8, high: 10}],
+      [{text: '[-8, 10]'}, {guesstimateType: 'NORMAL', low: -8, high: 10}],
       [{text: '[8, 10]', guesstimateType: 'UNIFORM'}, {guesstimateType: 'UNIFORM', low: 8, high: 10}], [{text: '[8,10]', guesstimateType: 'NORMAL'}, {guesstimateType: 'NORMAL', low: 8, high: 10}],
       [{text: '[-5, 5]', guesstimateType: 'NORMAL'}, {guesstimateType: 'NORMAL', low: -5, high: 5}],
       [{text: '[0, 5]', guesstimateType: 'NORMAL'}, {guesstimateType: 'NORMAL', low: 0, high: 5}],
