+++ src/lib/guesstimator/samplers/simulator-worker/tests/Simulator-test.js
@@ -1,7 +1,9 @@
-import {requiresManySamples} from '../simulator/evaluator.js';
+import {requiresManySamples, Evaluate} from '../simulator/evaluator.js';
 import {expect} from 'chai';
 
-describe('Function Sampler', () => {
+const n = 5000
+
+describe('Simulator', () => {
   describe('#requiresManySamples', () => {
     const examples = [
       [{ text: '3', inputs: {}}, false],
@@ -24,5 +26,32 @@ describe('Function Sampler', () => {
         expect(requiresManySamples(inputs.text, inputs.inputs)).to.equal(e[1])
       })
     }).map(e => e())
+  })
+
+  describe('#evaluate', () => {
+    it('a function with no inputs', () => {
+      const samples = Evaluate('34', n, [])
+      expect(samples.values).to.deep.eq([34])
+    })
+
+    describe('function guesstimateType', () => {
+      const examples = [
+        ['3', {}, 1, 3],
+        ['3*AK', {AK: [3]}, 1, 9],
+        ['(3*AK)+BA', {AK: [3], BA: [2]}, 1, 11],
+      ]
+
+      examples.map(e => () => {
+        it(`guesstimate ${e[0]} with value count ${e[1]} has correct number of values`, () => {
+          const _sample = Evaluate(e[0], e[2], e[1])
+          expect(_sample.values.length).to.equal(e[2])
+        })
+
+        it(`guesstimate ${e[0]} with value count ${e[1]} has correct first value`, () => {
+          const _sample = Evaluate(e[0], e[2], e[1])
+          expect(_sample.values[0]).to.equal(e[3])
+        })
+      }).map(e => e())
+    })
   })
 })
